Customize: Move create_pathlist and path_is_library to a utils module
