+++ spyder_kernels/customize/spyderpdb.py
@@ -3,131 +3,40 @@
 #
 # Licensed under the terms of the MIT License
 # (see spyder_kernels/__init__.py for details)
-# -----------------------------------------------------------------------------
-#
-# IMPORTANT NOTE: Don't add a coding line here! It's not necessary for
-# site files
-#
-# Spyder debugger
-#
+
+"""Spyder debugger."""
+
 import bdb
-import os
 import pdb
-import re
 import sys
-import sysconfig
 import logging
 import traceback
 
 from IPython.core.getipython import get_ipython
 from IPython.core.debugger import Pdb as ipyPdb
 
-from spyder_kernels.py3compat import TimeoutError, PY2, _print, isidentifier
 from spyder_kernels.comms.frontendcomm import CommError, _frontend_request
+from spyder_kernels.customize.utils import path_is_library
+from spyder_kernels.py3compat import TimeoutError, PY2, _print, isidentifier
 
 if not PY2:
     from IPython.core.inputtransformer2 import TransformerManager
     basestring = (str,)
 else:
-    from IPython.core.inputsplitter import IPythonInputSplitter as\
-        TransformerManager
+    from IPython.core.inputsplitter import IPythonInputSplitter as TransformerManager
 
 
 logger = logging.getLogger(__name__)
 
 
-def create_pathlist():
-    """
-    Create list of Python library paths to be skipped from module
-    reloading and Pdb steps.
-    """
-    # Get standard installation paths
-    try:
-        paths = sysconfig.get_paths()
-        standard_paths = [paths['stdlib'],
-                          paths['purelib'],
-                          paths['scripts'],
-                          paths['data']]
-    except Exception:
-        standard_paths = []
-
-    # Get user installation path
-    # See spyder-ide/spyder#8776
-    try:
-        import site
-        if getattr(site, 'getusersitepackages', False):
-            # Virtualenvs don't have this function but
-            # conda envs do
-            user_path = [site.getusersitepackages()]
-        elif getattr(site, 'USER_SITE', False):
-            # However, it seems virtualenvs have this
-            # constant
-            user_path = [site.USER_SITE]
-        else:
-            user_path = []
-    except Exception:
-        user_path = []
-
-    return standard_paths + user_path
-
-
-def path_is_library(path, initial_pathlist=None):
-    """Decide if a path is in user code or a library according to its path."""
-    # Compute DEFAULT_PATHLIST only once and make it global to reuse it
-    # in any future call of this function.
-    if 'DEFAULT_PATHLIST' not in globals():
-        global DEFAULT_PATHLIST
-        DEFAULT_PATHLIST = create_pathlist()
-
-    if initial_pathlist is None:
-        initial_pathlist = []
-
-    pathlist = initial_pathlist + DEFAULT_PATHLIST
-
-    if path is None:
-        # Path probably comes from a C module that is statically linked
-        # into the interpreter. There is no way to know its path, so we
-        # choose to ignore it.
-        return True
-    elif any([p in path for p in pathlist]):
-        # We don't want to consider paths that belong to the standard
-        # library or installed to site-packages.
-        return True
-    elif not os.name == 'nt':
-        # Paths containing the strings below can be part of the default
-        # Linux installation, Homebrew or the user site-packages in a
-        # virtualenv.
-        patterns = [
-            r'^/usr/lib.*',
-            r'^/usr/local/lib.*',
-            r'^/usr/.*/dist-packages/.*',
-            r'^/home/.*/.local/lib.*',
-            r'^/Library/.*',
-            r'^/Users/.*/Library/.*',
-            r'^/Users/.*/.local/.*',
-        ]
-
-        if [p for p in patterns if re.search(p, path)]:
-            return True
-        else:
-            return False
-    else:
-        return False
-
-
-# =============================================================================
-# Pdb adjustments
-# =============================================================================
 class SpyderPdb(ipyPdb, object):  # Inherits `object` to call super() in PY2
     """
     Extends Pdb to add features:
 
-     - Process ipython magic.
+     - Process IPython magics.
      - Accepts multiline input.
-     - Option to process 'pre_execute' and 'post_execute' events.
      - Better interrupt signal handling.
      - Option to skip libraries while stepping.
-     - Break into threads.
      - Add completion to non-command code.
     """
 
@@ -316,7 +225,6 @@ class SpyderPdb(ipyPdb, object):  # Inherits `object` to call super() in PY2
             self.pdb_ignore_lib = pdb_settings['pdb_ignore_lib']
             if self.starting:
                 self.set_spyder_breakpoints(pdb_settings['breakpoints'])
-
         except (CommError, TimeoutError):
             logger.debug("Could not get breakpoints from the frontend.")
