+++ spyder_kernels/customize/tests/test_spydercustomize.py
@@ -17,7 +17,7 @@ import pytest
 
 # Local imports
 from spyder_kernels.customize.spydercustomize import UserModuleReloader
-from spyder_kernels.customize.spyderpdb import create_pathlist
+from spyder_kernels.customize.utils import create_pathlist
 from spyder_kernels.py3compat import to_text_string
